Clean up, fixes and update build script

+++ tools/build_all_platforms.py
@@ -27,7 +27,8 @@ android_buildir = "TMP_build_android"
 mac_buildir = "TMP_build_mac"
 windows_buildir = "TMP_build_windows"
 emscripten_buildir = "TMP_build_emscripten"
-output_dir = "distrib-" + version
+output_dir = "/home/gee/art/superflu_riteurnz/distrib-" + version
+steam_dir = output_dir + "/steam"
 appname = gamename + "-" + version
 
 packages_to_do = set()
@@ -70,21 +71,13 @@ def chdir(folder):
     else:
         os.chdir(folder)
 
-def folder_size(folder):
-    size = 0
-    for path, dirs, files in os.walk(folder):
-        for f in files:
-            fp = os.path.join(path, f)
-            size += os.lstat(fp).st_size
-    print("SIZE = " + str(size))
-    return size // (1024 * 1024)
-
 all_begin = time.perf_counter()
 
 print("### CLEANING")
 begin = time.perf_counter()
 run_cmd("rm -rf " + output_dir)
 run_cmd("mkdir " + output_dir)
+run_cmd("mkdir " + steam_dir)
 end = time.perf_counter()
 print("  -> done in " + str(int(end - begin)) + "s\n")
 
@@ -115,7 +108,7 @@ if todo("Data"):
     chdir(data_dir)
     copy_data_dir = appname + "-data"
     run_cmd("cp -r " + data_folder + " " + copy_data_dir)
-    run_cmd("zip -r ../" + output_dir + "/" + appname + "-data.zip " + copy_data_dir)
+    run_cmd("zip -r " + output_dir + "/" + appname + "-data.zip " + copy_data_dir)
     chdir("..")
     run_cmd("rm -rf " + data_dir)
     end = time.perf_counter()
@@ -130,8 +123,24 @@ if todo("Linux"):
     run_cmd("cmake -DCMAKE_BUILD_TYPE=Release -DSOSAGE_DATA_FOLDER=" + data_folder + " ..")
     run_cmd("make -j 6")
     run_cmd("cpack")
-    run_cmd("cp *.deb ../" + output_dir + "/" + appname + "-gnunux.deb")
-    run_cmd("cp *.rpm ../" + output_dir + "/" + appname + "-gnunux.rpm")
+    run_cmd("cp *.deb " + output_dir + "/" + appname + "-gnunux.deb")
+    run_cmd("cp *.rpm " + output_dir + "/" + appname + "-gnunux.rpm")
+    chdir("..")
+    run_cmd("rm -rf " + linux_buildir)
+    end = time.perf_counter()
+    print("  -> done in " + str(int(end - begin)) + "s\n")
+
+if todo("Steam"):
+    print("### BUILDING LINUX/STEAMOS")
+    begin = time.perf_counter()
+    run_cmd("rm -rf " + linux_buildir)
+    run_cmd("mkdir " + linux_buildir)
+    chdir(linux_buildir)
+    run_cmd("mkdir install")
+    run_cmd('schroot --chroot steamrt_scout_amd64 -- sh -c "CC=/usr/bin/gcc-9 CXX=/usr/bin/g++-9 cmake -DCMAKE_BUILD_TYPE=Release -DSOSAGE_DATA_FOLDER=' + data_folder + ' -DYAML_INCLUDE_DIR=' + libyaml_source_path + '/include/ -DLZ4_INCLUDE_DIR=' + lz4_source_path + '/lib/ -DCMAKE_INSTALL_PREFIX=./install .."')
+    run_cmd('schroot --chroot steamrt_scout_amd64 -- sh -c "make -j 6"')
+    run_cmd('schroot --chroot steamrt_scout_amd64 -- sh -c "make install"')
+    run_cmd("cp -r install " + steam_dir + "/linux")
     chdir("..")
     run_cmd("rm -rf " + linux_buildir)
     end = time.perf_counter()
@@ -147,9 +156,53 @@ if todo("Appimage"):
     chdir("appimg")
     run_cmd("cp ../../platform/appimage/create_appimage.sh .")
     run_cmd("bash create_appimage.sh " + linuxdeploy + " " + data_folder)
-    run_cmd("cp *.AppImage ../../" + output_dir + "/" + appname + "-gnunux.AppImage")
+    run_cmd("cp *.AppImage " + output_dir + "/" + appname + "-gnunux.AppImage")
     chdir("../..")
     run_cmd("rm -rf " + appimg_buildir)
+    end = time.perf_counter()
+    print("  -> done in " + str(int(end - begin)) + "s\n")
+
+if todo("Mac"):
+    print("### BUILDING MAC")
+    begin = time.perf_counter()
+    run_cmd("rm -rf " + mac_buildir)
+    run_cmd("mkdir " + mac_buildir)
+    chdir(mac_buildir)
+    run_cmd("cmake -DCMAKE_BUILD_TYPE=Release -DSOSAGE_DATA_FOLDER=" + data_folder
+            + " -DCMAKE_TOOLCHAIN_FILE=../cmake/Toolchain-osxcross.cmake"
+            + " -DSDL2_INCLUDE_DIR=" + mac_sdl_folder
+            + " -DCMAKE_INSTALL_PREFIX=./install ..")
+    run_cmd("make -j 6")
+    run_cmd("make install")
+    run_cmd("python3 ../tools/fix_mac_lib_paths.py install/" + gamename + ".app/Contents/libs/*.dylib")
+    run_cmd("python3 ../tools/fix_mac_lib_paths.py install/" + gamename + ".app/Contents/MacOS/" + gamename)
+    run_cmd("cp -r install " + steam_dir + "/mac")
+    run_cmd("genisoimage -V " + gamename + ".app -D -R -apple -no-pad -o " + gamename + ".dmg install")
+    run_cmd("cp " + gamename + ".dmg " + output_dir + "/" + appname + "-mac.dmg")
+    chdir("..")
+    run_cmd("rm -rf " + mac_buildir)
+    end = time.perf_counter()
+    print("  -> done in " + str(int(end - begin)) + "s\n")
+
+if todo("Windows"):
+    print("### BUILDING WINDOWS")
+    begin = time.perf_counter()
+    run_cmd("rm -rf " + windows_buildir)
+    run_cmd("mkdir " + windows_buildir)
+    chdir(windows_buildir)
+    run_cmd("cmake -DCMAKE_BUILD_TYPE=Release -DSOSAGE_DATA_FOLDER=" + data_folder
+            + " -DCMAKE_TOOLCHAIN_FILE=../cmake/Toolchain-mingw32.cmake "
+            + "-DSDL2_INCLUDE_DIR=" + windows_sdl_folder + " ..")
+    run_cmd("make -j 6")
+    run_cmd("cpack")
+    run_cmd("cp *-win32.exe " + output_dir + "/" + appname + "-windows.exe")
+    run_cmd("cmake -DCMAKE_INSTALL_PREFIX=./install .")
+    run_cmd("mkdir install")
+    run_cmd("make -j 6")
+    run_cmd("make install")
+    run_cmd("cp -r install " + steam_dir + "/windows")
+    chdir("..")
+    run_cmd("rm -rf " + windows_buildir)
     end = time.perf_counter()
     print("  -> done in " + str(int(end - begin)) + "s\n")
 
@@ -166,62 +219,11 @@ if todo("Android"):
             + " -DSDL2_TTF_SOURCE_PATH:PATH=" + sdl2_ttf_source_path
             + " -DYAML_SOURCE_PATH:PATH=" + libyaml_source_path
             + " -DLZ4_SOURCE_PATH:PATH=" + lz4_source_path + " ..")
-    chdir(android_buildir + "/android")
+    chdir("android")
     run_cmd("./gradlew assembleDebug --parallel --max-workers=6")
-    run_cmd("cp app/build/outputs/apk/debug/app-debug.apk ../../" + output_dir + "/" + appname + "-android.apk")
+    run_cmd("cp app/build/outputs/apk/debug/app-debug.apk " + output_dir + "/" + appname + "-android.apk")
     chdir("../..")
     run_cmd("rm -rf " + android_buildir)
-    end = time.perf_counter()
-    print("  -> done in " + str(int(end - begin)) + "s\n")
-
-if todo("Windows"):
-    print("### BUILDING WINDOWS")
-    begin = time.perf_counter()
-    run_cmd("rm -rf " + windows_buildir)
-    run_cmd("mkdir " + windows_buildir)
-    chdir(windows_buildir)
-    run_cmd("cmake -DCMAKE_BUILD_TYPE=Release -DSOSAGE_DATA_FOLDER=" + data_folder
-            + " -DCMAKE_TOOLCHAIN_FILE=../cmake/Toolchain-mingw32.cmake "
-            + "-DSDL2_INCLUDE_DIR=" + windows_sdl_folder + " ..")
-    run_cmd("make -j 6")
-    run_cmd("cpack")
-    run_cmd("cp *-win32.exe ../" + output_dir + "/" + appname + "-windows.exe")
-    chdir("..")
-    run_cmd("rm -rf " + windows_buildir)
-    end = time.perf_counter()
-    print("  -> done in " + str(int(end - begin)) + "s\n")
-
-if todo("Mac"):
-    print("### BUILDING MAC")
-    begin = time.perf_counter()
-    run_cmd("rm -rf " + mac_buildir)
-    run_cmd("mkdir " + mac_buildir)
-    chdir(mac_buildir)
-    run_cmd("cmake -DCMAKE_BUILD_TYPE=Release -DSOSAGE_DATA_FOLDER=" + data_folder
-            + " -DCMAKE_TOOLCHAIN_FILE=../cmake/Toolchain-osxcross.cmake"
-            + " -DSDL2_INCLUDE_DIR=" + mac_sdl_folder
-            + " -DCMAKE_INSTALL_PREFIX=./install ..")
-    run_cmd("make -j 6")
-    run_cmd("make install")
-    run_cmd("dd if=/dev/zero of=" + gamename + ".dmg bs=1M count=" + str(16 + folder_size('install/' + gamename)) + " status=progress")
-    run_cmd("mkfs.hfsplus -v " + gamename + ".app " + gamename + ".dmg")
-    run_cmd("mkdir mnt")
-    run_cmd("sudo mount -o loop " + gamename + ".dmg mnt")
-    run_cmd("python3 ../tools/fix_mac_lib_paths.py install/" + gamename + "/Contents/Frameworks/*.dylib")
-    run_cmd("python3 ../tools/fix_mac_lib_paths.py install/" + gamename + "/Contents/MacOS/" + gamename)
-    chdir("install")
-    run_cmd("mv " + gamename + " " + gamename + ".app")
-    run_cmd("sudo cp -a -r " + gamename + ".app ../mnt/")
-    chdir("..")
-    run_cmd("sudo cp -a " + raw_data_folder + "/resources/Info.plist mnt/" + gamename + ".app/Contents/")
-    run_cmd("sudo cp -a " + raw_data_folder + "/resources/icon.icns mnt/" + gamename + ".app/Contents/Resources/")
-#    run_cmd("sudo mkdir mnt/.background/")
-#    run_cmd("sudo cp -a " + raw_data_folder + "/resources/background.png mnt/.background/")
-#    run_cmd("sudo cp -a " + raw_data_folder + "/resources/Applications mnt/")
-    run_cmd("sudo umount mnt")
-    run_cmd("cp " + gamename + ".dmg ../" + output_dir + "/" + appname + "-mac.dmg")
-    chdir("..")
-    run_cmd("rm -rf " + mac_buildir)
     end = time.perf_counter()
     print("  -> done in " + str(int(end - begin)) + "s\n")
 
@@ -235,9 +237,9 @@ if todo("Browser"):
             + " -DYAML_SOURCE_PATH:PATH=" + libyaml_source_path
             + " -DLZ4_SOURCE_PATH:PATH=" + lz4_source_path + " ..")
     run_cmd("make -j 6")
-    run_cmd("mkdir ../" + output_dir + "/" + appname + "-web/")
-    run_cmd("cp -r data ../" + output_dir + "/" + appname + "-web/")
-    run_cmd("cp *.data *.js *.wasm ../" + output_dir + "/" + appname + "-web/")
+    run_cmd("mkdir " + output_dir + "/" + appname + "-web/")
+    run_cmd("cp -r data " + output_dir + "/" + appname + "-web/")
+    run_cmd("cp *.data *.js *.wasm " + output_dir + "/" + appname + "-web/")
     chdir("..")
     run_cmd("rm -rf " + emscripten_buildir)
     end = time.perf_counter()
